+++ data/src/main/java/com/saegil/data/remote/FeedServiceImpl.kt
@@ -3,9 +3,11 @@ package com.saegil.data.remote
 
 import android.util.Log
 import com.saegil.data.model.NoticeListDto
-import io.ktor.client.*
+import io.ktor.client.HttpClient
 import io.ktor.client.call.body
-import io.ktor.client.request.*
+import io.ktor.client.request.accept
+import io.ktor.client.request.get
+import io.ktor.client.request.headers
 import io.ktor.http.ContentType
 import io.ktor.http.URLBuilder
 import javax.inject.Inject
@@ -18,7 +20,7 @@ class FeedServiceImpl @Inject constructor(
         query: String?,
         sourceId: String?,
         size: Int?,
-        lastId: Int?
+        lastId: Long?
     ): NoticeListDto? {
         val urlBuilder = URLBuilder(HttpRoutes.NOTICES).apply {
             query?.let { parameters.append("query", it) }
